datatables 2

+++ model/acta.php
@@ -2,21 +2,25 @@
 
 require_once 'conection.php';
 
-class Acta {
+class Acta
+{
 
-    public static function toJson($name, $data) {
+    public static function toJson($name, $data)
+    {
         header('Content-type:application/json;charset=utf-8');
         return json_encode($data, JSON_UNESCAPED_SLASHES);
     }
 
-    public static function toJson2($name, $data){
+    public static function toJson2($name, $data)
+    {
         header('Content-type:application/json;charset=utf-8');
         return json_encode([
             $name => $data
         ]);
     }
 
-    public static function getAll() {
+    public static function getAll()
+    {
         $table = "acta";
         $sql = "SELECT * FROM {$table}";
         $stmt = Conection::conect()->prepare($sql);
@@ -27,9 +31,10 @@ class Acta {
         return $result;
     }
 
-    public static function getBy($colum, $value) {
+    public static function getBy($colum, $value)
+    {
         $table = "acta";
-        if(is_string($value)){
+        if (is_string($value)) {
             $value = '"' . $value . '"';
         }
 
@@ -40,7 +45,8 @@ class Acta {
         return $stmt->fetchAll();
     }
 
-    public static function getBetween($initialDate, $endDate) {
+    public static function getBetween($initialDate, $endDate)
+    {
         $table = "acta";
 
         $initialDate = '"' . $initialDate . '"';
@@ -55,7 +61,8 @@ class Acta {
         return $result;
     }
 
-    public static function getPending() {
+    public static function getPending()
+    {
         $table = "acta";
         $now = date("Y-m-d");
         $now = '"' . $now . '"';
@@ -69,7 +76,8 @@ class Acta {
         return $result;
     }
 
-    public static function save($data) {
+    public static function save($data)
+    {
         $table = "acta";
         $sql = "INSERT INTO {$table} (creator_id, issue, created_date, start_time, end_time, in_charge, order_of_day, facts_description) VALUES (?,?,?,?,?,?,?,?)";
         $stmt = Conection::conect()->prepare($sql);
@@ -77,7 +85,7 @@ class Acta {
         $now = date("Y-m-d");
         $stmt->bindParam(1, $data['new_creator_id']);
         $stmt->bindParam(2, $data['new_issue']);
-        $stmt->bindParam(3, $data['new_start_date']);
+        $stmt->bindParam(3, $data['new_created_date']);
         $stmt->bindParam(4, $data['new_start_time']);
         $stmt->bindParam(5, $data['new_end_time']);
         $stmt->bindParam(6, $data['new_in_charge']);
@@ -90,7 +98,8 @@ class Acta {
         }
     }
 
-    public static function update($data) {
+    public static function update($data)
+    {
         $table = "acta";
         $sql = "UPDATE {$table} SET issue = ?, start_time = ?, end_time = ?, in_charge = ?, order_of_day = ?, facts_description = ? WHERE id = ?";
         $stmt = Conection::conect()->prepare($sql);
@@ -101,7 +110,7 @@ class Acta {
         $stmt->bindParam(4, $data['up_in_charge']);
         $stmt->bindParam(5, $data['up_order_of_day']);
         $stmt->bindParam(6, $data['up_facts_description']);
-        $stmt->bindParam(7, $data['id_acta']);
+        $stmt->bindParam(7, $data['up_id_acta']);
 
         if ($stmt->execute()) {
             return "ok";
@@ -110,7 +119,8 @@ class Acta {
         }
     }
 
-    public static function updateValue($column1, $value1, $column2, $value2) {
+    public static function updateValue($column1, $value1, $column2, $value2)
+    {
         $table = "users";
         $stmt = Conection::conect()->prepare("UPDATE $table SET $column1 = ? WHERE $column2 = ?");
 
@@ -124,16 +134,19 @@ class Acta {
         }
     }
 
-    public static function delete($id) {
+    public static function delete($id)
+    {
         $table = "acta";
-        $sql = "DELETE FROM {$table} WHERE id = ?";
-        $stmt = Conection::conect()->prepare($sql);
-        $stmt->bindParam(1, $id);
+        $result = self::getBy("id", $id);
+        if ($result != null) {
+            $sql = "DELETE FROM {$table} WHERE id = ?";
 
-        if ($stmt->execute()) {
+            $stmt = Conection::conect()->prepare($sql);
+            $stmt->bindParam(1, $id);
+
+            $stmt->execute();
             return "ok";
-        } else {
-            return "error";
         }
+        return "error";
     }
 }
